Migrate `local-firebase` server to using local Firebase emulator
Previously, the `local-firebase` server deployed an in-memory server
which mirrored its subscriptions to the Firestore database via
`FirebaseSubscriptionMirror`.

Now, it deploys a local web server, similar to `appengine-web`, which
uses a local Firebase emulator. This way, no real credentials are used
and no real Firebase database is involved.

The dependency on `FirebaseSubscriptionMirror` is gone.